Bound Hecate test subprocesses
Add an explicit timeout to Hecate CLI invocations used by architecture
helper tests and wrap timeout failures with contextual errors.

Move helper-unit coverage into a focused module and document the
fixture-only star import suppression.
