+++ tests/fixtures/architecture/api_imports_star_reexported_outbound_adapter/api.py
@@ -1,5 +1,5 @@
 """Violating inbound adapter that star-imports a re-exported outbound adapter."""
 
-from . import *  # noqa: F403
+from . import *  # noqa: F403  # Fixture requires star-imported re-exports.
 
 ADAPTER = StorageAdapter()  # noqa: F405
